(fix) add form launcher validation for HEI forms
